+++ README.md
@@ -3,7 +3,15 @@
 [![CircleCI](https://circleci.com/gh/nerves-project/nerves_ssh/tree/main.svg?style=svg)](https://circleci.com/gh/nerves-project/nerves_ssh/tree/main)
 [![Hex version](https://img.shields.io/hexpm/v/nerves_ssh.svg "Hex version")](https://hex.pm/packages/nerves_ssh)
 
-Manage an SSH daemon and its subsystems on Nerves devices
+Manage an SSH daemon and its subsystems on Nerves devices. It has the following
+features:
+
+1. Automatic startup of the SSH daemon on initialization
+2. Ability to hook the SSH daemon into a supervision tree of your choosing
+3. Easy setup of SSH firmware updates for Nerves
+4. Easy shell and exec setup for Erlang and Elixir
+5. Some protection from easy-to-make mistakes that would cause ssh to not be
+   available
 
 ## Usage
 
@@ -15,10 +23,17 @@ Most importantly, it makes it possible to segment failures in other OTP
 applications from terminating the daemon and it recovers from rare scenarios
 where the daemon terminates without automatically restarting.
 
+It can be started automatically as an OTP application or hooked into a
+supervision tree of your creation. Most Nerves users start it automatically as
+an OTP application. This is easy, but may be limiting and it requires that you
+use the application environment. See the following sections for options:
+
+### Starting as an OTP application
+
 If you're using [`:nerves_pack`](https://hex.pm/packages/nerves_pack) v0.4.0 or
-later, you don't need to do anything except, perhaps, modify the `:nerves_ssh`'s
-configuration in your `config.exs`. If you are not using `:nerves_pack`, add
-`:nerves_ssh` to your `mix` dependency list:
+later, you don't need to do anything except verify the `:nerves_ssh`
+configuration in your `config.exs` (see below). If you are not using
+`:nerves_pack`, add `:nerves_ssh` to your `mix` dependency list:
 
 ```elixir
 def deps do
@@ -39,9 +54,27 @@ config :shoehorn,
 your main OTP application stops, OTP may stop `:nerves_ssh`, and that would make
 your device inaccessible via SSH.
 
+### Starting as part of one of your supervision trees
+
+If you want to do this, make sure that you do NOT specify `:nerves_ssh` in your
+`config.exs`. The `:nerves_ssh` key decides whether or not to automatically launch
+based on this.
+
+Then when specifying the children for your supervisor, add `NervesSSH` like
+this:
+
+```elixir
+    {NervesSSH, nerves_ssh_options}
+```
+
+The `nerves_ssh_options` should be a `NervesSSH.Options` struct. See the
+`Configuration` section option fields that you may specify. Calling
+`NervesSSH.Options.with_defaults(my_options_list)` to build the
+`nerves_ssh_options` value is one way of getting reasonable defaults.
+
 ## Configuration
 
-NervesSSH supports a few pieces of configuration via the application config:
+NervesSSH supports the following configuration items:
 
 * `:authorized_keys` - a list of SSH authorized key file string
 * `:user_passwords` - a list of username/password tuples (stored in the
@@ -64,12 +97,12 @@ getting started. As needs become more sophisticated, you can pass options to
 
 ### Public key authentication
 
-Public ssh keys can be included in the `config.exs` so that matching clients can
-connect. These come from files like your `~/.ssh/id_rsa.pub` or
-`~/.ssh/id_ecdsa.pub` that were created when you created your `ssh` keys. If you
-haven't done this, the following
+Public ssh keys can be specified so that matching clients can connect. These
+come from files like your `~/.ssh/id_rsa.pub` or `~/.ssh/id_ecdsa.pub` that were
+created when you created your `ssh` keys. If you haven't done this, the
+following
 [article](https://help.github.com/articles/generating-a-new-ssh-key-and-adding-it-to-the-ssh-agent/)
-may be helpful. Here's an example:
+may be helpful. Here's an example that uses the application config:
 
 ```elixir
 config :nerves_ssh,
